Removed dependency on ~/.ssh/id_rsa.
There's now a mock ssh key in spec/poolparty. Key is reopened and the spec dir is appended to the list of possible locations for a key to appear. If you have a key named id_rsa in ~/.ssh it will still be found first, most likely, but is no longer required for rake spec to run successfully.